cosc562os

Module sbi

COSC562 Fall 2026 Operating System

Model’s the SBI call functionality. The SBI can be called to execute ecall instructions with elevated permissions.

Jason Weeks - jweeks12 September 13th, 2026

Macros

sbicall 🔒

SBI call: runs the RISC-V SBI, which is handled by SBI firmware. Inputs:

Structs

WakeFrame

Stores planned startup values for a secondary hart.

Enums

HartState

Enumerator to represent the state of a hart

SbiError

Enumeration to abstract the SBI error code.

SbiResult

SbiResult result. Either returns the error, or signals no error.

Functions

debug_write_char

Uses an SBI call to write a byte to the kernel output.

get_spec_version

Get Specification versions

hart_start

Start a hart using an SBI call.

hart_status

Check the specification of a hart.

hart_stop

Stop a hart using an SBI call.

hart_suspend

Makes an SBI call to stop a hart. Generally used to enter a low power state while remaining responsive.

reboot

reboot the system using an SBI call (function id 2)

shutdown

power the system off using an SBI call (function id 0)